Rename to ContentApiController
