+++ src/Umbraco.Cms.Api.Content/Controllers/ByIdContentApiController.cs
@@ -1,7 +1,5 @@
-using System.Net;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
-using Umbraco.Cms.Core;
 using Umbraco.Cms.Core.ContentApi;
 using Umbraco.Cms.Core.Models.ContentApi;
 using Umbraco.Cms.Core.Models.PublishedContent;
@@ -9,25 +7,25 @@ using Umbraco.Cms.Core.PublishedCache;
 
 namespace Umbraco.Cms.Api.Content.Controllers;
 
-public class ByRouteContentController : ContentApiControllerBase
+public class ByIdContentApiController : ContentApiControllerBase
 {
     private readonly IApiContentBuilder _apiContentBuilder;
 
-    public ByRouteContentController(IPublishedSnapshotAccessor publishedSnapshotAccessor, IApiContentBuilder apiContentBuilder)
+    public ByIdContentApiController(IPublishedSnapshotAccessor publishedSnapshotAccessor, IApiContentBuilder apiContentBuilder)
         : base(publishedSnapshotAccessor)
         => _apiContentBuilder = apiContentBuilder;
 
     /// <summary>
-    ///     Gets a content item by route.
+    ///     Gets a content item by id.
     /// </summary>
-    /// <param name="url">The path to the content item.</param>
+    /// <param name="id">The unique identifier of the content item.</param>
     /// <returns>The content item or not found result.</returns>
-    [HttpGet("{url}")]
+    [HttpGet("{id:guid}")]
     [MapToApiVersion("1.0")]
     [ProducesResponseType(typeof(IApiContent), StatusCodes.Status200OK)]
     [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status400BadRequest)]
     [ProducesResponseType(typeof(NotFoundResult), StatusCodes.Status404NotFound)]
-    public async Task<IActionResult> ByRoute(string url)
+    public async Task<IActionResult> ById(Guid id)
     {
         IPublishedContentCache? contentCache = GetContentCache();
 
@@ -36,9 +34,7 @@ public class ByRouteContentController : ContentApiControllerBase
             return await Task.FromResult(BadRequest(ContentCacheNotFoundProblemDetails()));
         }
 
-        var decodedPath = string.Format("/{0}", WebUtility.UrlDecode(url).TrimStart(Constants.CharArrays.ForwardSlash));
-
-        IPublishedContent? contentItem = contentCache.GetByRoute(decodedPath);
+        IPublishedContent? contentItem = contentCache.GetById(id);
 
         if (contentItem is null)
         {
