+++ src/Umbraco.Cms.Api.Content/Controllers/ByRouteContentApiController.cs
@@ -1,5 +1,7 @@
+using System.Net;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using Umbraco.Cms.Core;
 using Umbraco.Cms.Core.ContentApi;
 using Umbraco.Cms.Core.Models.ContentApi;
 using Umbraco.Cms.Core.Models.PublishedContent;
@@ -7,25 +9,25 @@ using Umbraco.Cms.Core.PublishedCache;
 
 namespace Umbraco.Cms.Api.Content.Controllers;
 
-public class ByIdContentController : ContentApiControllerBase
+public class ByRouteContentApiController : ContentApiControllerBase
 {
     private readonly IApiContentBuilder _apiContentBuilder;
 
-    public ByIdContentController(IPublishedSnapshotAccessor publishedSnapshotAccessor, IApiContentBuilder apiContentBuilder)
+    public ByRouteContentApiController(IPublishedSnapshotAccessor publishedSnapshotAccessor, IApiContentBuilder apiContentBuilder)
         : base(publishedSnapshotAccessor)
         => _apiContentBuilder = apiContentBuilder;
 
     /// <summary>
-    ///     Gets a content item by id.
+    ///     Gets a content item by route.
     /// </summary>
-    /// <param name="id">The unique identifier of the content item.</param>
+    /// <param name="url">The path to the content item.</param>
     /// <returns>The content item or not found result.</returns>
-    [HttpGet("{id:guid}")]
+    [HttpGet("{url}")]
     [MapToApiVersion("1.0")]
     [ProducesResponseType(typeof(IApiContent), StatusCodes.Status200OK)]
     [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status400BadRequest)]
     [ProducesResponseType(typeof(NotFoundResult), StatusCodes.Status404NotFound)]
-    public async Task<IActionResult> ById(Guid id)
+    public async Task<IActionResult> ByRoute(string url)
     {
         IPublishedContentCache? contentCache = GetContentCache();
 
@@ -34,7 +36,9 @@ public class ByIdContentController : ContentApiControllerBase
             return await Task.FromResult(BadRequest(ContentCacheNotFoundProblemDetails()));
         }
 
-        IPublishedContent? contentItem = contentCache.GetById(id);
+        var decodedPath = string.Format("/{0}", WebUtility.UrlDecode(url).TrimStart(Constants.CharArrays.ForwardSlash));
+
+        IPublishedContent? contentItem = contentCache.GetByRoute(decodedPath);
 
         if (contentItem is null)
         {
